Cope with ssh transfer delays over 2 minutes or with system clock changes.
Any of these changes would causes excessive ssh delays. In particular
if a transfer took more than 2 minutes elapsed after the previous
transfer the calculated delay would end up being negative which was
then cast to an unsigned making a delay of around 2 billion milliseconds.

Fix the delay by ensuring that the new delay is relative to the current
time and also aligned with the same seconds offset into a whole minute
as was originally present.